pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["kongms002 <ms020221@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask-sqlalchemy = "^3.1.1"
mongoengine = "^0.28.1"
flasgger = "^0.9.7.1"
flask-swagger = "^0.2.14"
python-dotenv = "^1.0.1"
flask-jwt-extended = "^4.6.0"
pydantic = "^2.6.3"
datetime = "^5.4"
flask-cors = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
